+++ scripts/llm/performance/pretrain_llama3_8b.py
@@ -17,23 +17,23 @@ from typing import Optional
 import nemo_run as run
 from utils import get_comm_overlap_callback_idx, hf_tokenizer, parse_cli_args, slurm_executor
 
-from nemo.collections.llm.recipes.llama31_405b import pretrain_recipe
+from nemo.collections.llm.recipes.llama3_8b import pretrain_recipe
 from nemo.collections.llm.recipes.precision.mixed_precision import bf16_with_fp8_mixed
 from nemo.lightning.pytorch.callbacks.garbage_collection import GarbageCollectionCallback
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
 
-NUM_NODES = 72
+NUM_NODES = 1
 NUM_GPUS_PER_NODE = 8
 MICRO_BATCH_SIZE = 1
-GLOBAL_BATCH_SIZE = 252
-TP_SIZE = 8
-PP_SIZE = 9
+GLOBAL_BATCH_SIZE = 128
+TP_SIZE = 1
+PP_SIZE = 1
 CP_SIZE = 2
-VP_SIZE = 7
+VP_SIZE = None
 MAX_STEPS = 100
 
 
-def llama3_405b_performance_recipe(
+def llama3_8b_performance_recipe(
     compute_dtype: str,
     num_nodes: int,
     num_gpus_per_node: int,
@@ -46,7 +46,7 @@ def llama3_405b_performance_recipe(
     max_steps: int,
 ):
     """
-    llama3 405b pre-train recipe aimed at achieving best possible performance.
+    llama3 8b pre-train recipe aimed at achieving best possible performance.
 
     NOTE: Use fp8 precision training with caution. It might not give desirable results.
     """
@@ -56,7 +56,7 @@ def llama3_405b_performance_recipe(
     recipe.data.micro_batch_size = mbs
     recipe.data.global_batch_size = gbs
     recipe.data.num_train_samples = max_steps * gbs * mbs  # ensure only 1 epoch for whole run
-    recipe.data.tokenizer = hf_tokenizer("meta-llama/Llama-3.1-405B")
+    recipe.data.tokenizer = hf_tokenizer("meta-llama/Meta-Llama-3-8B")
 
     recipe.trainer.max_steps = max_steps
     recipe.trainer.num_nodes = num_nodes
@@ -77,9 +77,6 @@ def llama3_405b_performance_recipe(
     # compute dtype configs
     if compute_dtype.lower() == "fp8":
         recipe.trainer.plugins = bf16_with_fp8_mixed()
-        recipe.trainer.callbacks[comm_overlap_callback_idx].tp_comm_overlap_cfg.proj_fprop.fp8_buf = True
-        recipe.trainer.callbacks[comm_overlap_callback_idx].tp_comm_overlap_cfg.fc2_fprop.fp8_buf = True
-
     recipe.trainer.plugins.grad_reduce_in_fp32 = False  # bf16 grad dtype
 
     # callback configs
@@ -112,7 +109,7 @@ if __name__ == "__main__":
 
     exp_name = "_".join(
         [
-            f"llama3_405b",
+            f"llama3_8b",
             args.compute_dtype,
             f"{NUM_NODES}nodes",
             f"tp{TP_SIZE}_pp{PP_SIZE}_cp{CP_SIZE}_vp{VP_SIZE}",
@@ -130,10 +127,11 @@ if __name__ == "__main__":
         args.container_image,
         custom_mounts=[],
         custom_env_vars={},
-        retries=0,
+        hf_token=args.hf_token,
+        nemo_home=args.nemo_home,
     )
 
-    recipe = llama3_405b_performance_recipe(
+    recipe = llama3_8b_performance_recipe(
         args.compute_dtype,
         NUM_NODES,
         NUM_GPUS_PER_NODE,
@@ -154,7 +152,7 @@ if __name__ == "__main__":
         # following line ensures file is at- `<log_dir>/lightning_logs/tb_logs/default/<tfevents_file>`
         recipe.log.log_dir = "/nemo_run/lightning_logs"
 
-    plugins = [PerfEnvPlugin(enable_vboost=True, nccl_pp_comm_chunksize=2097152)]
+    plugins = [PerfEnvPlugin(enable_vboost=True)]
     if args.enable_profiling:
         plugins.append(NsysPlugin(start_step=5, end_step=6))
